Converted context-menu to Cordova plugin
