--- pkg/bb10/ext/manifest.json
@@ -1,6 +1,0 @@
-{
-    "global": true,
-    "namespace": "blackberry.bbui",
-    "dependencies": []
-}
-
