+++ pkg/bb10/ext/plugin.xml
@@ -1,0 +1,38 @@
+<!--
+ Licensed to the Apache Software Foundation (ASF) under one
+ or more contributor license agreements.  See the NOTICE file
+ distributed with this work for additional information
+ regarding copyright ownership.  The ASF licenses this file
+ to you under the Apache License, Version 2.0 (the
+ "License"); you may not use this file except in compliance
+ with the License.  You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing,
+ software distributed under the License is distributed on an
+ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ KIND, either express or implied.  See the License for the
+ specific language governing permissions and limitations
+ under the License.
+-->
+<plugin xmlns="http://cordova.apache.org/ns/plugins/1.0"
+    xmlns:rim="http://www.blackberry.com/ns/widgets"
+    id="com.blackberry.bbui"
+    version="1.0.0">
+
+    <description>BlackBerry 10 BBui APIs</description>
+    <license>Apache 2.0</license>
+
+    <js-module src="www/client.js">
+        <clobbers target="blackberry.bbui" />
+    </js-module>
+
+    <platform name="blackberry10">
+        <source-file src="src/blackberry10/index.js" />
+        <config-file target="www/config.xml" parent="/widget">
+            <feature name="com.blackberry.bbui" value="com.blackberry.bbui" />
+        </config-file>
+    </platform>
+
+</plugin>
